This piece of work is a follow up to the earlier #68810 where we provided a Patches endpoint for CTCT. This story alters the Elevation value returned from the endpoint from a float (4 bytes) to ushort (2 bytes), in an effort to reduce the overall response size sent back to the machine.

- The value returned by 3DP is also incremented by 1 for all non null values.
- Calculation of the elevation offset is now required to be done by the API consumer.

More work may follow as we further refine the Protobuf response behaviour.

Related work items: #68810

+++ src/WebApiModels/ProductionData/ResultHandling/PatchCellHeightResult.cs
@@ -17,12 +17,12 @@ namespace VSS.Productivity3D.WebApi.Models.ProductionData.ResultHandling
     /// </summary>
     [JsonProperty(PropertyName = "elevation")]
     [ProtoMember(1)]
-    public float Elevation { get; private set; }
+    public ushort Elevation { get; private set; }
 
     /// <summary>
     /// Static constructor.
     /// </summary>
-    public static PatchCellHeightResult Create(float elevation)
+    public static PatchCellHeightResult Create(ushort elevation)
     {
       return new PatchCellHeightResult
       {
